+++ frontend/src/App.js
@@ -40,7 +40,8 @@ import PrintBarcode from "./components/printBarcode/Index";
 import NonConformIndex from "./components/nonconform/index";
 import SampleBatchEntrySetup from "./components/batchOrderEntry/SampleBatchEntrySetup.js";
 import AuditTrailReportIndex from "./components/reports/auditTrailReport/Index.js";
-import DictionaryManagement from "./components/Dictionary/DictionaryManagement.js";
+import DictionaryManagement from "./components/admin/masterListsPage/Dictionary/DictionaryManagement.js";
+import MasterListsPage from "./components/admin/masterListsPage/index.js";
 
 export default function App() {
   let i18nConfig = {
@@ -228,6 +229,18 @@ export default function App() {
                   role="Global Administrator"
                 />
                 <SecureRoute
+                  path="/admin/MasterListsPage"
+                  exact
+                  component={() => <MasterListsPage />}
+                  role="Global Administrator"
+                />
+                <SecureRoute
+                  path="/DictionaryManagement"
+                  exact
+                  component={() => <DictionaryManagement />}
+                  role="Global Administrator"
+                />
+                <SecureRoute
                   path="/PathologyDashboard"
                   exact
                   component={() => <PathologyDashboard />}
@@ -320,12 +333,6 @@ export default function App() {
                   path="/PrintBarcode"
                   exact
                   component={() => <PrintBarcode />}
-                  role="Reception"
-                />
-                <SecureRoute
-                  path="/DictionaryManagement"
-                  exact
-                  component={() => <DictionaryManagement />}
                   role="Reception"
                 />
                 <SecureRoute
