+++ frontend/src/components/addOrder/add-order.scss
@@ -41,7 +41,7 @@
 }
 
 .orderLegendBody {
-  background-color: #ffffff;
+  background-color: #FFFFFF;
   border-radius: 5px;
   margin: 2%;
   padding: 2%;
@@ -52,6 +52,40 @@
     margin: 0;
     padding: 0;
   }
+
+.adminMasterListsPage {
+  background-color: #FFFFFF;
+  border-radius: 5px;
+  margin: 2%;
+  padding: 2%;
+}
+
+@media (max-width: 1024px) {
+  .orderWorkFlowDiv {
+    margin: 0;
+    padding: 0;
+  }
+}
+
+.adminMasterListsPage {
+  background-color: #FFFFFF;
+  border-radius: 5px;
+  margin: 2%;
+  padding: 2%;
+}
+
+@media (max-width: 1024px) {
+  .orderWorkFlowDiv {
+    margin: 0;
+    padding: 0;
+  }
+}
+
+.adminMasterListsPage {
+  background-color: #FFFFFF;
+  border-radius: 5px;
+  margin: 2%;
+  padding: 2%;
 }
 
 .ProgressIndicator {
