+++ frontend/src/components/admin/masterListsPage/Dictionary/DictionaryManagement.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { FormattedMessage, injectIntl } from "react-intl";
-import "../Style.css";
+import "../../../Style.css";
 import { Heading, Grid, Column, Section } from "@carbon/react";
-import PageBreadCrumb from "../common/PageBreadCrumb";
+import PageBreadCrumb from "../../../common/PageBreadCrumb";
 
 const breadcrumbs = [
   { label: "home.label", link: "/" },
@@ -13,19 +13,7 @@ function DictionaryManagement() {
   return (
     <>
       <PageBreadCrumb breadcrumbs={breadcrumbs} />
-      <Grid fullWidth={true}>
-        <Column lg={16}>
-          <Section>
-            <Section>
-              <Heading>
-                <FormattedMessage id="dictionary.label.modify" />
-              </Heading>
-            </Section>
-          </Section>
-        </Column>
-      </Grid>
-      <br></br>
-      <div className="orderLegendBody">
+      <div className="adminMasterListsPage">
         <h1>Dictionary Management</h1>
       </div>
     </>
